CDI-392 move restriction from AfterDeploymentValidation to AfterBeanDiscovery
The whole sense of the Excpetion is to prevent random behaviour
which is caused by the classes being scanned in different order.
But it makes no sense to even prevent those methods in AfterBeanDiscovery.
The only thing we need to make sure in the containers is that any call to
AfterBeanDiscovery#addBean() etc will properly flush the internal
resolved-bean caches.