CDI-627 also re-enable the rules of CDI-1.0
This is for fixing a backward-incompatible change we introduced in CDI-1.1.

The whole reason for verifying the <class> entries is to ensure that a user
doesn't have typos or list classes which do not exists or are no
'candidates' for being an Alternative at least.
So it's all about preventing user errors. Disallowing to veto() @Alternative
classes (which was perfectly allowed in CDI-1.0) was overreaching.