remove App{Archive,ArchiveConfig,Deployment}, rename AppArchiveBuilder and ContextBuilder
The `AppArchive`, `AppArchiveConfig` and `AppDeployment` types
were redundant. Moreover, they looked like they can access whole
world, which prevents certain kinds of implementations. (Some
effort was made to design the API in these types so that whole
world access is not necessary, but it's a leaky abstraction.)
For those reasons, these types are removed altogether.

The `AppArchiveBuilder` type is renamed to `ScannedClasses`,
which is more descriptive (and shorter). The `addSubtypesOf`
method is removed, because it can't be implemented in certain
environments (notably, Portable Extensions).

The `ContextBuilder` type is renamed to `ContextConfig`, because
it isn't really used in the usual builder way and is quite close
to the other existing `*Config` types.

Additionally, some documentation is improved, especially for
types in the `@Discovery` phase.