add BeanConfigurator.priority
This is to allow setting a priority for synthetic beans created
using `BeanConfigurator`.

I briefly considered if `priority` should be added to
`BeanAttributesConfigurator` as well, but that would have
a ripple effect. It also doesn't reflect the present reality.
A custom `Bean` implementation can be added a priority
by implementing the `Prioritized` interface. Adding only
`BeanConfigurator.priority` is a perfect mirror of that
in the configurator API.