Allow @Enhancement to not restrict the types based on annotations
In case an unannotated type is added through `@Discovery`, such type
previously couldn't be transformed during `@Enhancement`. With this
commit, it is possible to specify that the set of types considered
for `@Enhancement` is not restricted based on annotations in any way.
To do that, the `@Enhancement#withAnnotations` member should be set
to an empty array.