Remove Enhancement.BeanDefiningAnnotations
The marker annotation `@Enhancement.BeanDefiningAnnotations`
used to represent all bean defining annotations for the purpose
of restricting the set of types on which `@Enhancement` should
be performed. That, however, is pretty useless: `@Enhancement`
is only performed on types that were discovered during type
discovery, and all such types have a bean defining annotation.
(Either they have it directly, or `@Dependent` is implied
in case of classes added through `@Discovery` that don't have
a bean defining annotation.)