fixes #567 split CDI-lite into own module
Technically there is no dependency from the 'Full' api to the lite api.
CDI-Lite is only an additional mechanism to support build time generation of proxies etc.
It is a kind of build-time replacement of the classic CDI Extension mechanism.
Servers should be free to support Full or Lite, but not forced to implement both.