Improve section numbering
CDI specification up to version 3.0 used to have the sections numbered
in a friendly way:

1. Architecture
2. Concepts
3. Programming model

And so on. With CDI 4.0 and the split to Lite and Full, sections were
renumbered in a more ugly fashion:

1. Architecture
2. CDI Lite
2.1. Concepts
2.2. Programming model
3. CDI Full

Note that this numbering is different from the TCK Audit tool expectations
and hence the TCK Audit reports are incorrect.

This commit restores the original numbering, which is more user-friendly
and also conforms to the expectations of the TCK Audit tool.