specification of standbys
Standbys are the opposite of alternatives: where alternatives take precedence
over non-alternatives, non-standbys take precedence over standbys. This pattern
has classically been known as "default beans", but that is a bad name (consists
of two words, both of them being used in the specification already).

In this commit, standbys can only be selected for application, they cannot be
selected just for a bean archive (in `beans.xml`). Perhaps this will have to
change, but present state should be good enough for now. Similarly, standby
producers currently do not inherit priority from their declaring beans, which
perhaps will have to change (for the sake of consistency), but present state
should be good enough for now.