fix the specification of alternatives
Certain parts of the specification said that alternatives are not available
for injection to classes in an archive, unless the archive is a bean archive.

This usually doesn't matter, because if an archive is not a bean archive,
beans in that archive are not discovered, but there's a set of "special"
classes (such as servlets) for which injection should still happen, even if
they are present in an archive that is not a bean archive.

This commit fixes that: alternative is available for injection into classes
in an archive if it is selected for the application, or if that archive
is a bean archive and the alternative is selected for that bean archive.