add @Eager
This annotation may be put on `@ApplicationScoped` beans to trigger eager
initialization. This occurs no later than at the time the container
fires the `Startup` event.

Putting the annotation on a bean of any other normal scope or on `@Dependent`
bean leads to definition error. Putting the annotation on a bean of any other
pseudo-scope leads to non-portable behavior (allowing implementations to
do the same for `@Singleton` beans).

It is possible to relax this in the future, if we find suitable meaning.
For example, eagerly initializing `@RequestScoped` beans typically
only makes sense for _some_ beans, depending on the nature of the request
(such as the URL in case of an HTTP request). My personal opinion is
that such conditions are best expressed in code, but I'm open to ideas.

`@Eager` also serves as a direct replacement of EJB `@Startup`, which also
only makes sense in case of application-wide beans (EJB `@Singleton`).