add @Eager
This annotation may be put on `@ApplicationScoped` beans to trigger eager
initialization. This occurs at the same time as the `Startup` event.

Putting the annotation on a bean of any other scope leads to definition
error. This may be relaxed in the future, if we find suitable meaning.

For example, eagerly initializing `@RequestScoped` beans typically
only makes sense for _some_ beans, depending on the nature of the request
(such as the URL in case of an HTTP request). My personal opinion is
that such conditions are best expressed in code, which is already
possible, because all beans can declare observer methods, which may
include arbitrary code.

`@Eager` is a shortcut for such observer in the most common scenario.
It also serves as a direct replacement of EJB `@Startup`, which also
only makes sense in case of application-wide beans (EJB `@Singleton`).