improve specification of InjectionPoint
The specification of `InjectionPoint` contains a small number of holes that
are arguably corner cases, but they should be specified nevertheless.

The most important one is dynamically obtained instance (`Instance`)
that is not injected. This can happen through several means:

- `BeanContainer.createInstance()`
- `CDI.current()`
- `SeContainer`

The specification didn't say what multiple methods should return in this case;
it didn't even explicitly say that an `InjectionPoint` should exist and indeed
it doesn't on one existing implementation. With this commit, the behavior is
clear: an `InjectionPoint` must exist and all methods besides `getType()` and
`getQualifiers()` return `null` (or `false`).

Further, it is not clear what the `getBean()` method should return in case
of a synthetic bean. Such injection points should, in portable extensions,
be created using `BeanManager.createInjectionPoint()`, but these methods
do not take a `Bean` and cannot figure out on their own which bean is
the injection point created for. So on all existing implementations that
support portable extensions, this method returns `null`. This is now
explicitly allowed.

Finally, this commit allows synthetic beans to have injection points that
do not correspond to a field or parameter. Currently, such injection points
should never exist, because injection points of synthetic beans should
be created using `BeanManager.createInjectionPoint()`, but it is not hard
to imagine people implementing `InjectionPoint` manually and returning
`null` from `getMember()` or `getAnnotated()`. We also plan to introduce
specifying injection points for synthetic beans in the build compatible
extensions API, where one is supposed to only specify the required type
and required qualifiers and it is entirely conceivable that no field or
parameter can possibly be determined upfront. So an `InjectionPoint`
of a synthetic bean is now explicitly allowed to return `null` from
the `getMember()` and `getAnnotated()` methods.